% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_plot.R
\name{multivariate_plot}
\alias{multivariate_plot}
\title{Plotting a multivariate graph}
\usage{
multivariate_plot(multi_output)
}
\arguments{
\item{multi_output}{an FactoMineR PCA analysis object}
}
\value{
a ggplot2 graphic object
}
\description{
This function takes a PCA analysis output from the FactoMineR R package
and plot the individuals' coordinates.
}
